+++ modules/core/dependency/python-ihm/ihm/dumper.py
@@ -805,7 +805,10 @@ class _ExternalReferenceDumper(Dumper):
                           "file_size_bytes", "details"]) as l:
             for r in self._ref_by_id:
                 repo = r.repo or self._local_files
-                file_path = self._posix_path(repo._get_full_path(r.path))
+                if r.path is None:
+                    file_path = None
+                else:
+                    file_path = self._posix_path(repo._get_full_path(r.path))
                 l.write(id=r._id, reference_id=repo._id,
                         file_path=file_path, content_type=r.content_type,
                         file_size_bytes=r.file_size, details=r.details)
@@ -1955,10 +1958,11 @@ class _CrossLinkDumper(Dumper):
                 seq1 = entity1.sequence
                 seq2 = entity2.sequence
                 pseudo = False
-                for np, p in enumerate((xl.pseudo1, xl.pseudo2)):
-                    if p:
+                for np, ps in enumerate((xl.pseudo1, xl.pseudo2)):
+                    if ps:
                         pseudo = True
-                        pseudo_xls.append((p, np, xl))
+                        for p in ps:
+                            pseudo_xls.append((p, np, xl))
                 l.write(id=xl._id, group_id=ex_xl._id,
                         entity_id_1=entity1._id, asym_id_1=xl.asym1._id,
                         seq_id_1=ex_xl.residue1.seq_id,